pyproject.toml
[tool.poetry]
name = "echo"
version = "0.1.0"
description = "Discord bot for gpt4all"
authors = ["minamorl <minamorl@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.0"
gpt4all = "^0.3.4"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
